requirements.txt
websocket-client>=0.57.0
requests>=2.22.0